feat: add GH workflow to auto-publish to gh-pages (#29)
This pull request introduces a new GitHub Actions workflow for deploying
the project to GitHub Pages and includes a version bump in the
`package.json` file. The most important changes are detailed below.

### Deployment Workflow Addition:
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R1-R42):
Added a new GitHub Actions workflow named "Deploy to GitHub Pages." This
workflow triggers on pushes to the `main` branch and includes steps for
checking out the repository, setting up Node.js, installing dependencies
with Yarn, building the project, creating a `CNAME` file, and deploying
the build directory to GitHub Pages.

### Version Update:
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L3-R3):
Incremented the project version from `0.11.1` to `0.11.2`.