fix: GitHub Action deploy workflow (#30)
This pull request updates the GitHub Actions workflow for deploying to
GitHub Pages. The changes aim to improve the workflow's reliability and
ensure a clean deployment process.

Workflow improvements:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L4-R13):
Changed the trigger for the deployment workflow from `push` to
`workflow_run`, making it dependent on the completion of the CI
workflow. Added a conditional to ensure deployment only occurs if the CI
workflow concludes successfully.

Deployment cleanup:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R46):
Added the `clean: true` parameter to the deployment job to ensure the
target directory is cleaned before publishing.